rename GridBoard widget file
